Mangle mcd for alternate layered update path
This is dirty but I needed the config drift check and validate on-disk
state to let me pivot to an image. The config itself should match the
pool but the OSImageURl obviously won't (and can't, because we build
after we render, so we have chicken + egg problems we'll have to sort
out somehow, but we knew that)

This uses a "newest-layered-image" annotation on the pool (tied to the
most recent image generated by the imagestream) , and a
"current-layered-image"/"desired-layered-image" to coordinate which
image the node should be using.

I have some controllers (sold separately) that I wrote using
controller-runtime that run outside the cluster and handle the
coordination of those annotations.