Port in external imagestream controller
I initially wrote an imagestream controller that lived outside
the cluster using controller-runtime, I am attempting to hack it into
the render controller the 'hard way' so I can watch imagestreams and
have the pool pull the new image.

This gets the pool to annotate itself with the newest image in the
imagestream. I'm not sure I like it, but it does work enough that by
hand (or via the node-controller) the annotation could be consumed to
set current/desired images