update.go: refactor calculatePostConfigChangeAction()
Currently, post config change actions are calculated as follows:
1) Reboot if there's a force file
2) Reboot if any MachineConfig fields change except for passwd
3) Calculate an action based on file changes

Layered updates will need to calculate post config change actions solely
based on files, as they will be diffing two images. 1) and 3) can be
consolidated into a single function that is shared between layered and
legacy updates. All MachineConfig changes (2) will result in file
changes on disk, which will cause a reboot if actions are calculated as
normal. This is correct for any change except for changes to passwd,
which currently do not require any post config change action. In order
to preserve this behavior, add a reboot file exception for ssh keys,
which is the only field the MCO supports modifying using passwd.