Add some more hacks to roll back later
Pushing this so I can reference it in my spike but everything outside of
render-controller is probably wrong and should be reverted aside from
the "ignore OSIMageURL parts of the config drift checker because
those have to be that way if we're rebasing to something that's
not the original OSImageURL.

This reworks things a little to add two annotations to the pool, one of
them is for the image like before, but the other one is the rendered
config that the image is equivalent to. This is ncecessary because we
can't control the timing of the image render, so the most recent image
from the imagestream isn't always THE image that matches the rendered
config.

We need to either delay the assignment of the pool.Spec.Config until
after the render or we need node-controller to wait until the image is
rendered to make the node an update candidate.