Make node_controller wait until image is rendered
To solve the race condition/coordination between when the rendered
config was rendered/assigned to the pool and when the image was
rendered, I just made the node_controller wait until the imagestream
rendered an image containing the rendered config (I parsed the image
labels to check).

This works all the way through and marks itself done, the pool doesn't
degrade, but the daemon doesn't have any of the reboot/postconfig logic
yet.