Fix dockerfile to work with butane
I just want to prove that this all works and it properly installs the
RPM, etc.

There is a magic file /etc/mco-butane.yaml that will:
- Get passed through to butane
- The resulting file will get ignition-liveapplied to the host and
- rpm-ostree ex rebuild will be run afterwards

This is not an honest attempt at a UX, it's just plumbing and proof that
it works. Someone could hotfix an image just by including the right
extensions in /etc/mco-butane.yaml.